Did Major Updates for the User Interface and User Options.
The Code still need to be optimized and more consolidated and cleaned. BUT everything is working good so far. I'd rewrite most of it in a new repos and make a separate class to handle everything for most of the tasks used here.

Here is the Updates:
- User Manager
-- Add/Edit Users
- Updated the Website Theme for Pages
- Added a Bar to show when you are in Admin Pages
- Added a Button that will auto show if you have perms for it.

Here is what I have left:
- Role Manager Page
-- Add/Remove Roles from Users
- Add a Remove/Delete User Function
- Convert the user.class.php functions to use a Global User Instance and if the UID is given then use that. Helps Eliminate the use of `$uid` in the code so much.